+++ mobile/android/android-components/components/browser/menu/src/test/java/mozilla/components/browser/menu/WebExtensionBrowserMenuBuilderTest.kt
@@ -4,11 +4,15 @@
 
 package mozilla.components.browser.menu
 
+import android.view.LayoutInflater
 import android.view.View
 import android.widget.ImageButton
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import mozilla.components.browser.menu.item.BackPressMenuItem
+import mozilla.components.browser.menu.item.BrowserMenuImageText
 import mozilla.components.browser.menu.item.WebExtensionBrowserMenuItem
+import mozilla.components.browser.menu.item.ParentBrowserMenuItem
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.WebExtensionState
 import mozilla.components.browser.state.store.BrowserStore
@@ -16,16 +20,39 @@ import mozilla.components.concept.engine.webextension.WebExtensionBrowserAction
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito
 
 @RunWith(AndroidJUnit4::class)
 class WebExtensionBrowserMenuBuilderTest {
 
     @Test
-    fun `web extension is added at the start when appendExtensionActionAtStart is true`() {
+    fun `WHEN there are no web extension actions THEN add-ons menu item invokes onAddonsManagerTapped`() {
+        var isAddonsManagerTapped = false
+        val store = BrowserStore()
+        val builder = WebExtensionBrowserMenuBuilder(
+            listOf(mockMenuItem(), mockMenuItem()),
+            store = store,
+            onAddonsManagerTapped = { isAddonsManagerTapped = true },
+            appendExtensionActionAtStart = true
+        )
+
+        val menu = builder.build(testContext)
+
+        val addonsManagerItem = menu.adapter.visibleItems[0] as? BrowserMenuImageText
+        val addonsManagerItemView =
+            LayoutInflater.from(testContext).inflate(addonsManagerItem!!.getLayoutResource(), null)
+        addonsManagerItem.bind(menu, addonsManagerItemView)
+        assertFalse(isAddonsManagerTapped)
+        addonsManagerItemView.performClick()
+        assertTrue(isAddonsManagerTapped)
+    }
+
+    @Test
+    fun `web extension sub menu add-ons manager sub menu item invokes onAddonsManagerTapped when clicked`() {
         val browserAction = WebExtensionBrowserAction("browser_action", false, mock(), "", 0, 0) {}
         val pageAction = WebExtensionBrowserAction("page_action", false, mock(), "", 0, 0) {}
 
@@ -40,22 +67,65 @@ class WebExtensionBrowserMenuBuilderTest {
             )
         )
 
-        val store = Mockito.spy(
-            BrowserStore(
-                BrowserState(
-                    extensions = extensions
-                )
+        val store = BrowserStore(
+            BrowserState(
+                extensions = extensions
             )
         )
 
-        val builder =
-            WebExtensionBrowserMenuBuilder(
-                listOf(mockMenuItem(), mockMenuItem()),
-                store = store,
-                appendExtensionActionAtStart = true)
+        var isAddonsManagerTapped = false
+        val builder = WebExtensionBrowserMenuBuilder(
+            listOf(mockMenuItem(), mockMenuItem()),
+            store = store,
+            onAddonsManagerTapped = { isAddonsManagerTapped = true },
+            appendExtensionActionAtStart = true
+        )
 
         val menu = builder.build(testContext)
 
+        val parentMenuItem = menu.adapter.visibleItems[0] as? ParentBrowserMenuItem
+        val subMenuItemSize = parentMenuItem!!.subMenu.adapter.visibleItems.size
+        assertEquals(6, subMenuItemSize)
+        val addOnsManagerMenuItem =
+            parentMenuItem.subMenu.adapter.visibleItems[subMenuItemSize - 1] as? BrowserMenuImageText
+        val addonsManagerItemView =
+            LayoutInflater.from(testContext).inflate(addOnsManagerMenuItem!!.getLayoutResource(), null)
+        addOnsManagerMenuItem.bind(menu, addonsManagerItemView)
+        assertFalse(isAddonsManagerTapped)
+        addonsManagerItemView.performClick()
+        assertTrue(isAddonsManagerTapped)
+        assertNotNull(addOnsManagerMenuItem)
+    }
+
+    @Test
+    fun `web extension submenu is added at the start when appendExtensionActionAtStart is true`() {
+        val browserAction = WebExtensionBrowserAction("browser_action", false, mock(), "", 0, 0) {}
+        val pageAction = WebExtensionBrowserAction("page_action", false, mock(), "", 0, 0) {}
+
+        val extensions = mapOf(
+            "id" to WebExtensionState(
+                "id",
+                "url",
+                "name",
+                true,
+                browserAction = browserAction,
+                pageAction = pageAction
+            )
+        )
+
+        val store = BrowserStore(
+            BrowserState(
+                extensions = extensions
+            )
+        )
+
+        val builder = WebExtensionBrowserMenuBuilder(
+            listOf(mockMenuItem(), mockMenuItem()),
+            store = store,
+            appendExtensionActionAtStart = true
+        )
+
+        val menu = builder.build(testContext)
         val anchor = ImageButton(testContext)
         val popup = menu.show(anchor)
 
@@ -64,17 +134,23 @@ class WebExtensionBrowserMenuBuilderTest {
 
         val recyclerAdapter = recyclerView.adapter!!
         assertNotNull(recyclerAdapter)
-        assertEquals(4, recyclerAdapter.itemCount)
+        assertEquals(3, recyclerAdapter.itemCount)
 
-        var menuItem = menu.adapter.visibleItems[0]
-        assertEquals("browser_action", (menuItem as WebExtensionBrowserMenuItem).action.title)
-
-        menuItem = menu.adapter.visibleItems[1]
-        assertEquals("page_action", (menuItem as WebExtensionBrowserMenuItem).action.title)
+        val parentMenuItem = menu.adapter.visibleItems[0] as? ParentBrowserMenuItem
+        val subMenuItemSize = parentMenuItem!!.subMenu.adapter.visibleItems.size
+        assertEquals(6, subMenuItemSize)
+        val backMenuItem = parentMenuItem.subMenu.adapter.visibleItems[0] as? BackPressMenuItem
+        val subMenuExtItemBrowserAction = parentMenuItem.subMenu.adapter.visibleItems[2] as? WebExtensionBrowserMenuItem
+        val subMenuExtItemPageAction = parentMenuItem.subMenu.adapter.visibleItems[3] as? WebExtensionBrowserMenuItem
+        val addOnsManagerMenuItem = parentMenuItem.subMenu.adapter.visibleItems[subMenuItemSize - 1] as? BrowserMenuImageText
+        assertNotNull(backMenuItem)
+        assertEquals("browser_action", subMenuExtItemBrowserAction!!.action.title)
+        assertEquals("page_action", subMenuExtItemPageAction!!.action.title)
+        assertNotNull(addOnsManagerMenuItem)
     }
 
     @Test
-    fun `web extension is added at the end when appendExtensionActionAtStart is false`() {
+    fun `web extension submenu is added at the end when appendExtensionActionAtStart is false`() {
         val browserAction = WebExtensionBrowserAction("browser_action", false, mock(), "", 0, 0) {}
         val pageAction = WebExtensionBrowserAction("page_action", false, mock(), "", 0, 0) {}
 
@@ -89,11 +165,9 @@ class WebExtensionBrowserMenuBuilderTest {
             )
         )
 
-        val store = Mockito.spy(
-            BrowserStore(
-                BrowserState(
-                    extensions = extensions
-                )
+        val store = BrowserStore(
+            BrowserState(
+                extensions = extensions
             )
         )
 
@@ -105,7 +179,6 @@ class WebExtensionBrowserMenuBuilderTest {
             )
 
         val menu = builder.build(testContext)
-
         val anchor = ImageButton(testContext)
         val popup = menu.show(anchor)
 
@@ -114,12 +187,19 @@ class WebExtensionBrowserMenuBuilderTest {
 
         val recyclerAdapter = recyclerView.adapter!!
         assertNotNull(recyclerAdapter)
-        assertEquals(4, recyclerAdapter.itemCount)
+        assertEquals(3, recyclerAdapter.itemCount)
 
-        var menuItem = menu.adapter.visibleItems[menu.adapter.itemCount - 1]
-        assertEquals("page_action", (menuItem as WebExtensionBrowserMenuItem).action.title)
-        menuItem = menu.adapter.visibleItems[menu.adapter.itemCount - 2]
-        assertEquals("browser_action", (menuItem as WebExtensionBrowserMenuItem).action.title)
+        val parentMenuItem = menu.adapter.visibleItems[recyclerAdapter.itemCount - 1] as? ParentBrowserMenuItem
+        val subMenuItemSize = parentMenuItem!!.subMenu.adapter.visibleItems.size
+        assertEquals(6, subMenuItemSize)
+        val backMenuItem = parentMenuItem.subMenu.adapter.visibleItems[0] as? BackPressMenuItem
+        val subMenuExtItemBrowserAction = parentMenuItem.subMenu.adapter.visibleItems[2] as? WebExtensionBrowserMenuItem
+        val subMenuExtItemPageAction = parentMenuItem.subMenu.adapter.visibleItems[3] as? WebExtensionBrowserMenuItem
+        val addOnsManagerMenuItem = parentMenuItem.subMenu.adapter.visibleItems[subMenuItemSize - 1] as? BrowserMenuImageText
+        assertNotNull(backMenuItem)
+        assertEquals("browser_action", subMenuExtItemBrowserAction!!.action.title)
+        assertEquals("page_action", subMenuExtItemPageAction!!.action.title)
+        assertNotNull(addOnsManagerMenuItem)
     }
 
     private fun mockMenuItem() = object : BrowserMenuItem {